Updating deployment scripts

+++ deployment/application_stop.sh
@@ -1,3 +1,10 @@
 #!/bin/bash
 PID=$(ps aux | grep /usr/bin/python3.5 | grep -v grep | awk '{print $2}')
-kill $PID
+
+if [ -z "$PID" ];
+then
+    exit 0
+else
+    kill $PID
+fi
+
